Move state of `facingMode` to CameraInput component to keep that state after Clear Photo Move SwitchFacingModeButton component to separate file
